+++ crates/matrix-sdk-ui/src/timeline/pagination.rs
@@ -129,7 +129,6 @@ impl super::Timeline {
                     }
                     RoomPaginationStatus::Paginating => {}
                 }
-
                 yield state;
             }
         });